+++ tests/testthat/test-apply.R
@@ -39,7 +39,7 @@ population <- createStudyPopulation(plpData,
                                     riskWindowEnd = 365,
                                     addExposureDaysToEnd = FALSE)
 
-index <- PatientLevelPrediction::personSplitter(population, test=0.2, seed=1)
+index <- PatientLevelPrediction::randomSplitter(population, test=0.2, seed=1)
 population <- merge(population, index)
 colnames(population)[colnames(population)=='index'] <- 'indexes'