+++ tests/testthat/test-dataSplitting.R
@@ -17,19 +17,19 @@
 library("testthat")
 context("Data splitting")
 
-test_that("Data splitting by person", {
+test_that("Data stratified splitting", {
 
   # error message checks
   population1 <- data.frame(rowId=1:20, outcomeCount=c(1,1,1,1,rep(0,16))) 
-  expect_error(personSplitter(population1, test=0.3, nfold=3))
+  expect_error(randomSplitter(population1, test=0.3, nfold=3))
   
   population2 <- data.frame(rowId=1:200, outcomeCount=c(rep(1,42),rep(0,158)))
-  expect_error(personSplitter(population2, test=0.3, nfold=-1))
-  expect_error(personSplitter(population2, test=1.5, nfold=5))
-  expect_error(personSplitter(population2, test=-1, nfold=5))
+  expect_error(randomSplitter(population2, test=0.3, nfold=-1))
+  expect_error(randomSplitter(population2, test=1.5, nfold=5))
+  expect_error(randomSplitter(population2, test=-1, nfold=5))
   
   # fold creation check 1 (fixed)
-  test <- personSplitter(population2, test=0.2, nfold=4)
+  test <- randomSplitter(population2, test=0.2, nfold=4)
   test <- merge(population2, test)
   test <- table(test$outcomeCount, test$index)
   test.returned <- paste(test, collapse='-')
@@ -39,7 +39,7 @@ test_that("Data splitting by person", {
   # fold creation check 2 (sum)
   size <- 500
   population3 <- data.frame(rowId=1:size, outcomeCount=c(rep(1,floor(size/3)),rep(0,size-floor(size/3)))) 
-  test <- personSplitter(population3, test=0.2, nfold=4) 
+  test <- randomSplitter(population3, test=0.2, nfold=4) 
   test <- merge(population3, test)
   test <- table(test$outcomeCount, test$index)
   expect_that(sum(test), equals(size))
@@ -49,7 +49,7 @@ test_that("Data splitting by person", {
   population4 <- data.frame(rowId=1:size,
                             outcomeCount=c(rep(1,floor(size/3)),
                                            rep(0,size-floor(size/3)))) 
-  test <- personSplitter(population4, test = 0.2, train = 0.4, nfold = 4)
+  test <- randomSplitter(population4, test = 0.2, train = 0.4, nfold = 4)
 
   tolerance = 5
   excludedPatients = 200
@@ -110,4 +110,36 @@ test_that("Data splitting by time", {
   expect_equal(length(test$index[test$index == 0]),
                excludedPatients)
   
+})
+
+
+
+test_that("Data splitting by subject", {
+# error message checks
+population1 <- data.frame(rowId=1:20, subjectId = 1:20, outcomeCount=c(1,1,1,1,rep(0,16))) 
+expect_error(subjectSplitter(population1, test=0.3, nfold=3))
+
+population2 <- data.frame(rowId=1:200,subjectId = 1:200, outcomeCount=c(rep(1,42),rep(0,158)))
+expect_error(subjectSplitter(population2, test=0.3, nfold=-1))
+expect_error(subjectSplitter(population2, test=1.5, nfold=5))
+expect_error(subjectSplitter(population2, test=-1, nfold=5))
+
+test <- subjectSplitter(population2, test=0.2, nfold=4)
+test <- merge(population2, test)
+test <- table(test$outcomeCount, test$index)
+test.returned <- paste(test, collapse='-')
+test.expected <- paste(matrix(c(32,32,32,31,31,8,9,9,8,8), ncol=5, byrow=T),collapse='-')
+expect_identical(test.returned, test.expected)
+
+# test that people are not in multiple folds
+population3 <- data.frame(rowId=1:200,subjectId = rep(1:50,4), outcomeCount=c(rep(1,42),rep(0,158)))
+test <- subjectSplitter(population3, test=0.2, nfold=3)
+test <- merge(population3, test)
+
+expect_equal(unique(table(test$subjectId[test$index==-1])), 4)
+expect_equal(unique(table(test$subjectId[test$index==2])), 4)
+expect_equal(unique(table(test$subjectId[test$index==3])), 4)
+expect_equal(unique(table(test$subjectId[test$index==1])), 4)
+
+
 })