+++ tests/testthat/test-plotting.R
@@ -44,7 +44,7 @@ lr_results <- tryCatch(runPlp(population = population, plpData = plpData,
                               modelSettings = lr_model, savePlpData = F,
                               savePlpResult = F, savePlpPlots =  F, 
                               saveEvaluation = F,
-                              testSplit='person', # this splits by person
+                              testSplit='stratified', # this splits by person
                               testFraction=0.25,
                               nfold=2))